C: Implement more efficient buffer resizing (#539)
## Problem

If the buffer is nearly full even one extra character can trigger an
expansion of the buffer. Since the buffer only grows by twice the number
of required characters per resize, in this case 2 characters, the buffer
has to be constantly resized.

### Visualization


![buffer_problem](https://github.com/user-attachments/assets/11658cf7-1ab5-41fe-8a75-fc6eaddfb80a)

## How the problem is addressed

Rather than just checking the required length, we test whether doubling
the current capacity will be enough. If it is, we expand to the doubled
capacity. If not, we double the required length itself and resize the
buffer to that size.

## Performance impact

Lexing a  [real world html page](https://shop.herthabsc.com)

Before: ~88.255ms

After: ~79.208ms

Parsing showed no significant performance impact though