C: Favor explicit buffer capacity over default capacity (#585)
This pull request is an alternative to #579 and reworks the buffer to
not have a default capacity anymore, but instead, let the caller decide
how big the initial buffer capacity should be.

This also allows callers to request enough capacity upfront if they know
the approximate or exact buffer buffer length, which then doesn't need
any buffer capacity expansions at a later point, thus removing the need
to reallocate.

Closes #579