C: Remove JSON Serialize Implementation (#586)
This pull request removes the JSON Serialize Implementation that we
haven't really made use of, so we are going to remove it for now.

If we end up needing it again, we can reference back to this pull
request and add the implementation back as it should be somewhat
straightforward to bring back.