C: Split lexer/parser alloc and init into separate steps (#513)
This PR changes the way lexers and parsers are initialized. 

Instead of making an allocation for the lexer/parser inside the init
function of the respective system, it allows the caller of the init
function to decide whether the lexer/parser data is going to live on the
stack or heap.

The lexer/parser lifetimes are limited to the scope of a single function
making it possible to use a stack variable.