Linter: Don't lint files with parser errors (#614)
This pull request updates the linter to not lint files with syntax
errors. This is to avoid false positives that might occur when trying to
lint a file that isn't properly parsed, which could also lead to a
"parser error" and "lint offense" diagnostic in the LSP.

Additionally, it updates the Linter test helpers to make sure all test
cases have valid syntax too. Since we have the `no-parser-errors` rule
we will still catch syntax errors, invalid syntax, and parser errors
with that rule in real applications.