Linter: Remove `html-no-space-in-tag` from default rules for now (#697)
The `html-no-space-in-tag` linter rule (#559 and #642) has quite a few
false positives and corrupts documents when using the `--fix` option
(see #695), which is why this pull request removes the
`html-no-space-in-tag` rule from the default rules for now.

We can enable this rule again in the future when we improve the accuracy
of the rule.